Added gradle and fixed bugs with borders.
This commit contains the version 0.1 of the theme
